==> Scripts/7.Cross-analysis.Rmd <==
---
  title: "cross_analysis"
author: "Miguel Ramon"
date: "`r Sys.Date()`"
output:
  html_document: default
pdf_document: default
---
  
  ################################## 

# CROSS-ANALYSIS

### PRIMATES' NEOPLASY PROJECT

### MIGUEL RAMON ALONSO

################################## 

```{r setup}

knitr::opts_knit$set(root.dir = "/home/miguel/TFM/Master_projects/NEOPLASY_PRIMATES")
knitr::opts_chunk$set(engine.path = list(
  python = '/home/miguel/TFM/Master_projects/NEOPLASY_PRIMATES/TreeCluster/venv/bin/python'))

```

```{r dirdef}

workingDir <- getwd()
dataDir <- file.path(workingDir, "Data")
resultsDir <- file.path(workingDir,"Out") 

```

```{r check-wd}

getwd()

```

### Seed and library loading

```{r,warning=FALSE,message=FALSE}

set.seed(1998)
library(dplyr)

```

# Common function

```{r common_functions}

# Function to read TSV files and return data frames
read_csv_to_df <- function(file) {
  df <- read.csv(file, sep = "\t")
  return(df)
}

createDir <- function(directory) {
  if (!file.exists(directory)) {
    dir.create(directory, recursive = TRUE)
  }
}

```

# Load filtered CAAS results

```{r caas_load}

# Filt CAAS: Directory path
filtered_caasDir <- file.path(resultsDir, "4.CAAS_analysis/CAASresults/14-11-run/Genes_tables/")

# Get a list of all TSV files in the directory and its subdirectories
filt_caas_files <- list.files(filtered_caasDir, pattern = "\\.tsv$", full.names = TRUE, recursive = FALSE)


# Use lapply to read all TSV files into a list of data frames
caas_dfs <- setNames(lapply(filt_caas_files, read_csv_to_df), tools::file_path_sans_ext(basename(filt_caas_files)))

```

# Load filtered RER results

```{r rer_load}

# Filt RER: Directory path
filtered_rerDir <- file.path(resultsDir, "7.RERConverge/RERConverge/RER_filtered_pval/rer_filtered_results.tsv")

rer_df <- read.csv(filtered_rerDir)

```

# Cross-check

```{r cross-check}

common_genes_data <- rer_df  # Initialize with rer_df or an empty data frame if needed
common_genes_list <- list()

cross_path <- file.path(resultsDir, "9.Cross-validation/")
createDir(cross_path)

for (name in names(caas_dfs)) {
  caas_df <- caas_dfs[[name]]
  
  common_genes_data <- caas_df
  
  common_genes_data <- common_genes_data %>%
    semi_join(rer_df, by = c("discovery.Gene" = "X")) %>%
    rename(
      Gene_symbol = discovery.Gene
    )
  
  cross_file <- paste0(cross_path, name, "_crossvalidated.tsv")
  write.csv(common_genes_data, cross_file, sep = "\t")  
  
  common_genes_list[[name]] <- common_genes_data
}
    
```
